Add a setting to let users pass additional arguments to shellcheck
This makes the plugin more flexible, as it does not necessarily need to be
updated if future versions of shellcheck add new command line arguments.

It also makes it possible for users to pass shellcheck that there currently
doesn't exist a specific configuration setting for (such as the
`--external-sources` flag).