Invoke shellcheck on individual files to overcome several existing limitations
Invoking shellcheck on each file individually has the following advantages:

  1. Allows the user to specify the exact set of files to invoke shellcheck on
     using regular Gradle syntax. This can be done by setting the new sourceFiles
     property instead of the original sources property.

  2. Because of the first item above, users are now able to run shellcheck
     against files that do not have a well-known extension (including files that
     have no extension).

  3. Better supports running shellcheck with --external-sources and
     --checked-sourced. In this particular use-case, the users typically want to
     include the main scripts that source other scripts, but exclude the other
     scripts (since they're already checked when the main scripts taht source them
     are checked).

	4. Avoids hitting a particular shell's maximum command line limit length (if
     there were a lot of sources).

There is a downside to this approach: it is slower than the original approach.
In particular, running shellcheck with docker is noticably slower. But the
benefits probably outweigh the downsides.

In the future, we could add a config setting to let users pass all the files to
a single shellcheck invocation (knowing full well that the shell's max command
length could be hit).